requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
mutagen==1.47.0
numpy==2.3.5
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
pydub==0.25.1
pygame==2.6.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pywin32-ctypes==0.2.3
setuptools==80.9.0
watchdog==6.0.0
